+++ .github/workflows/check-maven.yml
@@ -15,7 +15,7 @@ jobs:
           distribution: 'corretto'
           java-version: '8'
       - run: |-
-          cd Maven/OffScreen
+          cd Maven/Console
           mvn package
           cd ../JavaFX
           mvn package
