+++ Gradle/Console/build.gradle.kts
@@ -18,31 +18,33 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import static com.teamdev.jxbrowser.engine.RenderingMode.OFF_SCREEN;
+plugins {
+    // Apply the java plugin to add support for Java
+    java
 
-import com.teamdev.jxbrowser.browser.Browser;
-import com.teamdev.jxbrowser.engine.Engine;
+    // Apply the application plugin to add support for building a CLI application
+    application
 
-/**
- * This example demonstrates how to load a web page, wait until it is loaded
- * completely, and print its HTML without displaying any GUI.
- */
-public final class HelloOffScreen {
+    // Provides convenience methods for adding JxBrowser dependencies into a project
+    id("com.teamdev.jxbrowser") version "1.0.1"
+}
 
-    public static void main(String[] args) {
-        // Initialize Chromium.
-        Engine engine = Engine.newInstance(OFF_SCREEN);
+jxbrowser {
+    version = "7.36.1"
+}
 
-        // Create a Browser instance.
-        Browser browser = engine.newBrowser();
+dependencies {
+    // Use JxBrowser cross-platform binaries
+    implementation(jxbrowser.crossPlatform)
+}
 
-        // Load a web page and wait until it is loaded completely.
-        browser.navigation().loadUrlAndWait("https://html5test.com/");
+application {
+    // Define the main class for the application
+    mainClass.set("HelloConsole")
+}
 
-        // Print HTML of the loaded web page.
-        browser.mainFrame().ifPresent(frame -> System.out.println(frame.html()));
-
-        // Shutdown Chromium and release allocated resources.
-        engine.close();
-    }
+tasks.withType<JavaExec> {
+    // Assign all Java system properties from
+    // the command line to the JavaExec task.
+    systemProperties(System.getProperties().mapKeys { it.key as String })
 }
