+++ Gradle/Console/src/main/java/HelloConsole.java
@@ -18,4 +18,31 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-rootProject.name = "OffScreen"
+import static com.teamdev.jxbrowser.engine.RenderingMode.OFF_SCREEN;
+
+import com.teamdev.jxbrowser.browser.Browser;
+import com.teamdev.jxbrowser.engine.Engine;
+
+/**
+ * This example demonstrates how to load a web page, wait until it is loaded
+ * completely, and print its HTML without displaying any GUI.
+ */
+public final class HelloConsole {
+
+    public static void main(String[] args) {
+        // Initialize Chromium.
+        Engine engine = Engine.newInstance(OFF_SCREEN);
+
+        // Create a Browser instance.
+        Browser browser = engine.newBrowser();
+
+        // Load a web page and wait until it is loaded completely.
+        browser.navigation().loadUrlAndWait("https://html5test.com/");
+
+        // Print HTML of the loaded web page.
+        browser.mainFrame().ifPresent(frame -> System.out.println(frame.html()));
+
+        // Shutdown Chromium and release allocated resources.
+        engine.close();
+    }
+}
