+++ Maven/Console/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.teamdev.jxbrowser.quickstart.maven</groupId>
-    <artifactId>offscreen</artifactId>
+    <artifactId>console</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
